+++ Volume and Mesh Scattering/Halton.h
@@ -3,7 +3,7 @@
 #include <vector>
 #include <glm/glm.hpp>
 
-const float PI = 3.1415926535897932384626433832795;
+const double PI = 3.1415926535897932384626433832795;
 
 class halton
 {
